requirements.txt
matplotlib==1.5.2
pyserial==3.4